move exception to controller
