+++ src/main/java/com/kodilla/ecommercee/cart/controller/CartNotFoundException.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.cart.domain;
+package com.kodilla.ecommercee.cart.controller;
 
 public class CartNotFoundException extends RuntimeException {
     public CartNotFoundException(String message) {
